All of the departure constants got moved to voyager.py
